projet(style, refactor, perf): commenté le code + résoudre les problèmes de Dart Analysis + bonnes pratiques de flutter
