+++ lib/themes/string.dart
@@ -1,3 +1,4 @@
+// AppStrings class which stores all the strings
 abstract class AppStrings {
   static const appName = "App Renov Propriétaire";
   static const name = "Nom";
@@ -7,13 +8,17 @@ abstract class AppStrings {
   static const description = "Description";
   static const addDescription = "Ajouter une description ...";
   static const newCategory = "Nouvelle catégorie";
-  static const addNewCategory = "Nouvelle catégorie";
   static const generalPresentation = "Présentation générale";
   static const categoryWork = "Travaux de la catégorie";
+  static const categoryOfWork = "Catégorie de travaux";
   static const deleteCategory = "Supprimer la catégorie";
   static const modifCategory = "Modifier la catégorie";
   static const valider = "Valider";
-
+  static const catName = "Isolation thermique";
+  static const catPitch =
+      "Pour faire des économies d’énergie, et améliorer votre confort en hiver comme en été.";
+  static const catDesc =
+      "L’isolation de votre maison est primordiale. Une maison mal isolée est sujette à de sérieuses pertes énergétiques engendrant une perte de confort et d’argent.";
   static const selectElementToSeeMore =
       "Sélectionner un élément pour voir ses détails";
   static const areYouSureToDeleteCategory =
