+++ lib/view/widgets/category_item.dart
@@ -2,8 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:git/themes/app_theme.dart';
 import 'package:git/themes/dimension.dart';
-import 'package:git/widgets/app_icon.dart';
 
+import 'app_icon.dart';
+
+/*
+ * CategoryItem widget witch allows to select/unselect a category
+ *
+ * @params title, subtitle, onPressed, icon, selected
+ */
 class CategoryItem extends StatefulWidget {
   final String title;
   final String subtitle;
